+++ frontend/src/App.jsx
@@ -1,13 +1,7 @@
-import { useState } from 'react'
-import viteLogo from './assets/vite.svg'
-import heroImg from './assets/hero.png'
+import AppRoutes from "./routes/AppRoutes";
 
 function App() {
-  return (
-    <h1 className="text-4xl font-bold text-center mt-10">
-      DevTrack
-    </h1>
-  );
+  return <AppRoutes />;
 }
 
 export default App;